requirements.txt
openai>=1.0.0
pandas>=2.0.0
yfinance>=0.2.0
pyyaml>=6.0
